@@ -45,16 +45,16 @@
     "release.demo": "node --import tsx scripts/src/release-demo.mts"
   },
   "dependencies": {
-    "@angular/animations": "^17.0.6",
+    "@angular/animations": "^18.2.1",
     "@angular/cdk": "^17.0.3",
-    "@angular/common": "^17.0.6",
-    "@angular/compiler": "^17.0.6",
-    "@angular/core": "^17.0.6",
-    "@angular/forms": "^17.0.6",
+    "@angular/common": "^18.2.1",
+    "@angular/compiler": "^18.2.1",
+    "@angular/core": "^18.2.1",
+    "@angular/forms": "^18.2.1",
     "@angular/material": "^17.0.3",
-    "@angular/platform-browser": "^17.0.6",
-    "@angular/platform-browser-dynamic": "^17.0.6",
-    "@angular/router": "^17.0.6",
+    "@angular/platform-browser": "^18.2.1",
+    "@angular/platform-browser-dynamic": "^18.2.1",
+    "@angular/router": "^18.2.1",
     "@azure/msal-angular": "^3.0.4",
     "@azure/msal-browser": "^3.0.4",
     "@floating-ui/utils": "^0.1.4",
@@ -91,17 +91,17 @@
     "tslib": "^2.4.0",
     "typy": "^3.3.0",
     "windows-1252": "^3.0.4",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.6",
+    "@angular-devkit/build-angular": "^18.2.1",
     "@angular-eslint/builder": "^17.1.0",
     "@angular-eslint/eslint-plugin": "^17.1.0",
     "@angular-eslint/eslint-plugin-template": "^17.1.0",
     "@angular-eslint/schematics": "^17.1.0",
     "@angular-eslint/template-parser": "^17.1.0",
-    "@angular/cli": "^17.0.6",
-    "@angular/compiler-cli": "^17.0.6",
+    "@angular/cli": "^18.2.1",
+    "@angular/compiler-cli": "^18.2.1",
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@commitlint/format": "^19.3.0",
@@ -142,7 +142,7 @@
     "karma-jasmine-html-reporter": "~2.1.0",
     "lodash-es": "^4.17.21",
     "madge": "^6.1.0",
-    "ng-packagr": "^17.0.2",
+    "ng-packagr": "^18.2.1",
     "prettier": "^3.1.0",
     "sass": "^1.64.0",
     "semantic-release": "^23.0.0",
