@@ -1,4 +1,7 @@
-import { HttpClientModule } from '@angular/common/http';
+import {
+  provideHttpClient,
+  withInterceptorsFromDi
+} from '@angular/common/http';
 import { TestBed, inject } from '@angular/core/testing';
 
 import { TranslateModule } from '@ngx-translate/core';
@@ -8,7 +11,8 @@ import { LanguageService } from './language.service';
 describe('LanguageService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientModule, TranslateModule.forRoot()]
+      imports: [TranslateModule.forRoot()],
+      providers: [provideHttpClient(withInterceptorsFromDi())]
     });
   });
 
